requirement.txt
Flask==2.3.3
line-bot-sdk==3.5.0
openai>=1.0.0
python-dotenv==1.0.0
python-docx==1.1.0
PyPDF2==3.0.1
pandas==2.1.3
openpyxl==3.1.2
gunicorn==21.2.0
Werkzeug==2.3.7